+++ binutils-2.17/sim/bfin/dv-bfin_mmu.c
@@ -73,7 +73,7 @@ bfin_mmu_io_write_buffer (struct hw *me, const void *source,
 	     (int) nr_bytes, value));
 
   mmr_off = addr - mmu->base;
-  valuep = (void *)mmu + mmr_base() + mmr_off;
+  valuep = (void *)((unsigned long)mmu + mmr_base() + mmr_off);
 
   switch (mmr_off)
     {
@@ -119,7 +119,7 @@ bfin_mmu_io_read_buffer (struct hw *me, void *dest,
   HW_TRACE ((me, "read 0x%08lx length %d", (long) addr, (int) nr_bytes));
 
   mmr_off = addr - mmu->base;
-  valuep = (void *)mmu + mmr_base() + mmr_off;
+  valuep = (void *)((unsigned long)mmu + mmr_base() + mmr_off);
 
   switch (mmr_off)
     {
