+++ binutils-2.17/sim/bfin/dv-bfin_pll.c
@@ -59,7 +59,7 @@ bfin_pll_io_write_buffer (struct hw *me, const void *source,
 	     (int) nr_bytes, value));
 
   mmr_off = addr - pll->base;
-  valuep = (void *)pll + mmr_base() + mmr_off;
+  valuep = (void *)((unsigned long)pll + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -92,7 +92,7 @@ bfin_pll_io_read_buffer (struct hw *me, void *dest,
   HW_TRACE ((me, "read 0x%08lx length %d", (long) addr, (int) nr_bytes));
 
   mmr_off = addr - pll->base;
-  valuep = (void *)pll + mmr_base() + mmr_off;
+  valuep = (void *)((unsigned long)pll + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
